+++ configs/dataset/graph/NCI109.yaml
@@ -1,5 +1,5 @@
 loader:
-  _target_: topobenchmarkx.data.load.GraphLoader
+  _target_: topobenchmarkx.data.loaders.GraphLoader
   parameters: 
     data_domain: graph
     data_type: TUDataset
