+++ configs/dataset/graph/minesweeper.yaml
@@ -1,6 +1,6 @@
 # Dataset loader config
 loader:
-  _target_: topobenchmarkx.data.load.GraphLoader
+  _target_: topobenchmarkx.data.loaders.GraphLoader
   parameters: 
     data_domain: graph
     data_type: heterophilic
