+++ configs/dataset/graph/questions.yaml
@@ -1,5 +1,5 @@
 loader:
-  _target_: topobenchmarkx.data.load.GraphLoader
+  _target_: topobenchmarkx.data.loaders.GraphLoader
   parameters: 
     data_domain: graph
     data_type: heterophilic
