+++ configs/dataset/hypergraph/coauthorship_citeseer.yaml
@@ -1,4 +1,4 @@
-_target_: topobenchmarkx.data.load.GraphLoader
+_target_: topobenchmarkx.data.loaders.GraphLoader
 
 
 defaults:
