+++ configs/dataset/simplicial/karate_club.yaml
@@ -1,4 +1,4 @@
-_target_: topobenchmarkx.data.load.loaders.SimplicialLoader
+_target_: topobenchmarkx.data.loaders.SimplicialLoader
 
 # Data definition
 parameters:
